split contexts/meta to separate files
