+++ js/packages/web/src/utils/isValidHttpUrl.ts
@@ -1,0 +1,11 @@
+export function isValidHttpUrl(text: string) {
+  let url;
+
+  try {
+    url = new URL(text);
+  } catch (_) {
+    return false;
+  }
+
+  return url.protocol === 'http:' || url.protocol === 'https:';
+}
